Codigo/requirements.txt
absl-py==2.2.2
accelerate==0.34.2
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-cors==0.8.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
autogluon.common==1.2
autogluon.core==1.2
autogluon.features==1.2
autogluon.tabular==1.2
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
boto3==1.37.30
botocore==1.37.30
cachetools==5.5.2
catboost==1.2.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmdstanpy==1.2.5
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
coreforecast==0.0.12
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2025.3.2
fugue==0.9.1
future==1.0.0
gluonts==0.16.1
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
holidays==0.70
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
hyperopt==0.2.7
idna==3.10
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lightgbm==4.5.0
lightning==2.5.1
lightning-utilities==0.14.3
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.3
mlforecast==0.13.4
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
narwhals==1.34.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
neuralforecast==3.0.0
nixtla==0.6.6
notebook==7.4.0
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
opencensus==0.11.4
opencensus-context==0.1.3
optuna==4.2.1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pmdarima==2.0.4
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
prophet==1.1.6
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.1.1
pure_eval==0.2.3
py-spy==0.4.0
py4j==0.10.9.9
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==2.5.1
pytz==2025.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.39.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-base==0.12.2
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
sktime==0.36.1
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
stanio==0.5.1
statsforecast==1.7.8
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.1
toolz==0.12.1
torch==2.5.1
torchmetrics==1.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triad==0.9.8
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
utilsforecast==0.2.8
virtualenv==20.30.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
window_ops==0.0.15
wrapt==1.17.2
xgboost==2.1.4
yarl==1.19.0
zstandard==0.23.0
